changing the logic of backend to work with bigDecimal so it is more precise

+++ backend/src/main/java/com/Edu_App/services/AccountService.java
@@ -1,5 +1,6 @@
 package com.Edu_App.services;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.Edu_App.domain.entities.AccountEntity;
@@ -13,8 +14,8 @@ public interface AccountService {
     public List<AccountEntity> getAllAccounts();
     public List<AccountEntity> getAllActiveAccounts();
     public List<AccountEntity> findActiveAccountsByOwner(Integer userId);
-    public void depositInAccount(Integer id, double amount);
-    public void withdrawFromAccount(Integer id, double amount);
+    public void depositInAccount(Integer id, BigDecimal amount);
+    public void withdrawFromAccount(Integer id, BigDecimal amount);
     public void changeCurrency(Integer id, Integer newCurrencyId);
     public void deleteAccountById(Integer id);
     public String generateRandomIban();
