+++ backend/src/main/java/com/Edu_App/services/TransferService.java
@@ -1,5 +1,6 @@
 package com.Edu_App.services;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.Edu_App.domain.entities.TransferEntity;
@@ -12,6 +13,6 @@ public interface TransferService {
     public List<TransferEntity> getAllTransfersForUser(Integer userId);
     public List<TransferEntity> getTransfersSentByUser(Integer userId);
     public List<TransferEntity> getTransfersReceivedByUser(Integer userId);
-    public void validateTransferAmount(Integer senderId, double amount, Integer currencyId);
+    public void validateTransferAmount(Integer senderId, BigDecimal amount, Integer currencyId);
     public void executeTransfer(Integer transferId);
 }
